[HumanBenchmark][Backend][Libraries] refactor events endpoints
