+++ back-end/api/events.js
@@ -1,0 +1,166 @@
+const express = require('express');
+const { PrismaClient } = require('@prisma/client');
+const prisma = new PrismaClient();
+
+const events = express.Router();
+
+
+// [Post] /api/events
+events.post("/api/events", async (req, res) => {
+  // create new Calendar Event
+  const { title, description, start, end, type, createdById, participantIds } =
+    req.body;
+  try {
+    const event = await prisma.calendarEvent.create({
+      data: {
+        title,
+        description,
+        start: new Date(start),
+        end: new Date(end),
+        type,
+        createdBy: { connect: { id: createdById } },
+        participants: {
+          connect: participantIds.map((id) => ({ id })),
+        },
+      },
+    });
+
+    res.json(event);
+  } catch (err) {
+    console.log(err);
+  }
+});
+
+//[Put] /api/events/:id
+events.put("/api/events/:id", async (req, res) => {
+  //edit event
+  const { title, start, end, description, type, participantIds } = req.body;
+  const id = Number(req.params.id);
+
+  try {
+    const updated = await prisma.calendarEvent.update({
+      where: { id: id },
+      data: {
+        title,
+        description,
+        start: start ? new Date(start) : undefined,
+        end: end ? new Date(end) : undefined,
+        type,
+        participants: participantIds
+          ? { set: participantIds.map((id) => ({ id })) }
+          : undefined,
+      },
+    });
+
+    res.json(updated);
+  } catch (err) {
+    console.log(err);
+  }
+});
+
+//[Delete] /api/events/:id
+events.delete("/api/events/:id", async (req, res) => {
+  //Delete event
+  const id = Number(req.params.id);
+
+  try {
+    const deleted = await prisma.calendarEvent.delete({
+      where: { id: id },
+    });
+
+    res.json({ message: "Deleted", deleted });
+  } catch (err) {
+    console.log(err);
+  }
+});
+
+//[Get] /api/events/:userId
+events.get("/api/events/:userId", async (req, res) => {
+  // get User Calendar Events
+  const userId = Number(req.params.userId);
+
+  try {
+    const events = await prisma.calendarEvent.findMany({
+      where: {
+        participants: {
+          some: { id: userId },
+        },
+      },
+      include: {
+        participants: {
+          select: { id: true, username: true },
+        },
+        createdBy: {
+          select: { id: true, username: true },
+        },
+      },
+    });
+
+    res.json(events);
+  } catch (err) {
+    console.log(err);
+  }
+});
+
+//[Get] /api/availability/:userId
+events.get("/api/availability/:userId", async (req, res) => {
+  // get User Calendar Events
+  const userId = Number(req.params.userId);
+  try {
+    const availability = await prisma.availability.findMany({
+      where: { userId },
+    });
+
+    res.json(availability);
+  } catch (err) {}
+});
+
+//[Post] /api/availability
+events.post("/api/availability", async (req, res) => {
+  // create new availability block
+  const { userId, dayOfWeek, startTime, endTime } = req.body;
+
+  try {
+    const created = await prisma.availability.create({
+      data: { userId, dayOfWeek, startTime, endTime },
+    });
+
+    res.json(created);
+  } catch (err) {
+    console.log(err);
+  }
+});
+
+//[Put] /api/availability/:id
+events.put("/api/availability/:id", async (req, res) => {
+  // edit availability block
+  const id = Number(req.params.id);
+  const { startTime, endTime } = req.body;
+  try {
+    const updated = await prisma.availability.update({
+      where: { id: id },
+      data: { startTime, endTime },
+    });
+
+    res.json(updated);
+  } catch (err) {
+    console.log(err);
+  }
+});
+
+//[Delete] /api/availability/:id
+events.delete("/api/availability/:id", async (req, res) => {
+  const id = Number(req.params.id);
+
+  try {
+    const deleted = await prisma.availability.delete({
+      where: { id: id },
+    });
+
+    res.json({ message: "Availability deleted", deleted });
+  } catch (err) {
+    console.log(err);
+  }
+});
+
+module.exports = events;
