+++ back-end/package.json
@@ -25,7 +25,6 @@
     "express": "^5.1.0",
     "express-session": "^1.18.1",
     "helmet": "^8.1.0",
-    "moment": "^2.30.1",
     "prisma": "^6.10.1"
   }
 }
